sitemap: add seperate bongo file
