+++ go/src/socialapi/workers/sitemap/models/sitemapfile.go
@@ -26,34 +26,6 @@ type SitemapFile struct {
 	UpdatedAt time.Time
 }
 
-func NewSitemapFile() *SitemapFile {
-	return &SitemapFile{}
-}
-
-func (f SitemapFile) GetId() int64 {
-	return f.Id
-}
-
-func (f SitemapFile) BongoName() string {
-	return "sitemap.file"
-}
-
-func (f *SitemapFile) BeforeCreate() error {
-	if f.Name == "" {
-		return ErrNotSet
-	}
-
-	return nil
-}
-
-func (f *SitemapFile) Create() error {
-	return bongo.B.Create(f)
-}
-
-func (f *SitemapFile) One(q *bongo.Query) error {
-	return bongo.B.One(f, f, q)
-}
-
 func (f *SitemapFile) ByName(name string) error {
 	if name == "" {
 		return ErrNotSet
@@ -67,10 +39,6 @@ func (f *SitemapFile) ByName(name string) error {
 	return f.One(q)
 }
 
-func (f *SitemapFile) Update() error {
-	return bongo.B.Update(f)
-}
-
 func (f *SitemapFile) FetchAll() ([]SitemapFile, error) {
 	files := make([]SitemapFile, 0)
 
@@ -82,22 +50,10 @@ func (f *SitemapFile) FetchAll() ([]SitemapFile, error) {
 	return files, nil
 }
 
-func (f *SitemapFile) Some(data interface{}, q *bongo.Query) error {
-	return bongo.B.Some(f, data, q)
-}
-
-func (f *SitemapFile) ById(id int64) error {
-	return bongo.B.ById(f, id)
-}
-
 func (f *SitemapFile) UnmarshalBlob() (*ItemSet, error) {
 	set := NewItemSet()
 
 	return set, xml.Unmarshal(f.Blob, set)
-}
-
-func (f *SitemapFile) Delete() error {
-	return bongo.B.Delete(f)
 }
 
 func (f *SitemapFile) Purge() error {
